Refactor:PlaceInfoCard 관련 컴포넌트들 organisms 폴더로 이동
