+++ app/(pages)/(main)/record/search/page.tsx
@@ -1,5 +1,5 @@
 import ArrowBackTopBar from "@common/components/ui/topBar/ArrowBackTopBar/ArrowBackTopBar";
-import PlaceInfoCard from "@feature/place/components/PlaceInfo/molecules/PlaceInfoCard";
+import PlaceInfoCard from "@feature/place/components/PlaceInfo/organisms/PlaceInfoCard";
 import TextSearchBar from "@feature/search/components/SearchText/molecules/TextSearchBar";
 import { getTextSearchPlaceData } from "@feature/search/queries/getTextSearchPlaceData";
