+++ app/feature/curation/components/CurationDetail/molecules/CurationDetailInfoCard.tsx
@@ -2,7 +2,7 @@ import Image from "next/image";
 import { forwardRef, useState } from "react";
 import { CurationPlaceProps } from "@feature/curation/type";
 import LinkLayout from "@common/components/layout/Link/LinkLayout";
-import PlaceInfoCardBottom from "@feature/place/components/PlaceInfo/molecules/PlaceInfoCardBottom";
+import PlaceInfoCardBottom from "@feature/place/components/PlaceInfo/organisms/PlaceInfoCardBottom";
 import SliderLayout from "@common/components/layout/Slider/SliderLayout";
 import { useSetRecoilState } from "recoil";
 import { myCurationModalInfoSelector } from "@common/state/myCurationModal";
