+++ app/feature/place/components/PlaceScrapList/PlaceScrapList.tsx
@@ -1,7 +1,7 @@
 import GetPlaceScrapped from "@feature/place/queries/getPlaceScrapped";
 import dynamic from "next/dynamic";
 const PlaceInfoCard = dynamic(
-  () => import("../PlaceInfo/molecules/PlaceInfoCard")
+  () => import("../PlaceInfo/organisms/PlaceInfoCard")
 );
 
 export default async function PlaceScrapList() {
